talenta-backend/requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
pydantic[email]==2.6.4
httpx==0.27.0
python-multipart==0.0.9
